+++ block/src/woocommerce-tiered-price-table.js
@@ -1,0 +1,8 @@
+import TieredPriceTableCompatibility from './includes/TieredPriceTableCompatibility';
+
+window.addEventListener( 'DOMContentLoaded', () => {
+	const viewConfig = window.wtsViewObject || {
+		originalTaxDisplay: 'incl',
+	};
+	TieredPriceTableCompatibility.initialize( viewConfig.originalTaxDisplay );
+} );