+++ webpack.config.js
@@ -6,5 +6,10 @@ module.exports = {
 	entry: {
 		index: path.resolve( process.cwd(), 'block/src', 'index.js' ),
 		view: path.resolve( process.cwd(), 'block/src', 'view.js' ),
+		'woocommerce-tiered-price-table': path.resolve(
+			process.cwd(),
+			'block/src',
+			'woocommerce-tiered-price-table.js'
+		),
 	},
 };